requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-import-export==3.2.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.3
odfpy==1.4.1
openpyxl==3.1.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
tablib==3.5.0
uritemplate==4.1.1
urllib3==2.0.3
xlrd==2.0.1
xlwt==1.3.0
